Switch from long long to int64_t
git-svn-id: https://svn.mcs.anl.gov/repos/exm/sfw/lb/trunk@8560 dc4e9af1-7f46-4ead-bba6-71afc04862de

+++ code/src/adlb-defs.h
@@ -46,7 +46,7 @@ typedef enum
 /**
    Identifier for all ADLB data module user data
  */
-typedef long long adlb_datum_id;
+typedef int64_t adlb_datum_id;
 
 /**
    User data types
